SPSA gradient
